// src/types/core/preload.types.ts
export interface FileSystemNode {
  type: 'directory' | 'file';
  name: string;
  path: string;
  children?: FileSystemNode[];
}

// Represents data parsed from or to be serialized to YAML
export interface YamlData {
  matches?: Array<{
    trigger?: string; // Make optional as not all items might be matches
    replace?: string; // Make optional
    [key: string]: any;
  }>;
  [key: string]: any;
}